requirements.txt
aiosmtplib==4.0.2
app==0.0.1
boto3==1.39.12
fastapi==0.116.1
passlib==1.7.4
Pillow==11.3.0
pillow_heif==1.1.0
pydantic==2.11.7
python-multipart
python-dotenv==1.1.1
python_jose==3.5.0
SQLAlchemy==2.0.43
langchain
pypdf
langchain-community
python-docx
python-pptx
pandas
openpyxl
Pillow
pytesseract
sentence-transformers
chromadb
requests
transformers
torch
torchvision
accelerate
python-pptx
unstructured[pptx]

fastapi
uvicorn
python-multipart
pandas
boto3
pillow
pytesseract
transformers
langchain
langchain-community
langchain-core
langchain-groq
chromadb
sentence-transformers
huggingface-hub
unstructured
